// src/app/interfaces/playerInfo.ts
export interface PlayerInfo {
  id: number;
  name: string;
  appUserId: number;
  appUserName: string;
  appUserEmail: string;
  gameId: number;
  gameName: string;
  serverId: number;
  serverName: string;
  goalId: number;
  goalName: string;
  about: string;
  rankId: number;
  rankName: string;
  decency: number;
  teamId: number;
  teamName: string;
}